+++ packages/engine/src/audio/MediaModule.ts
@@ -23,7 +23,7 @@ All portions of the code written by the Infinite Reality Engine team are Copyrig
 Infinite Reality Engine. All Rights Reserved.
 */
 
-import { AudioAnalysisSystem } from '../audio/systems/AudioAnalysisSystem'
-import { MediaSystem } from '../audio/systems/MediaSystem'
+import { AudioAnalysisSystem } from './systems/AudioAnalysisSystem'
+import { MediaSystem } from './systems/MediaSystem'
 
 export { AudioAnalysisSystem, MediaSystem }
