+++ packages/ui/src/components/editor/properties/grab/index.tsx
@@ -29,32 +29,30 @@ CPAL-1.0 License
 import React, { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { isClient } from '@ir-engine/common/src/utils/getEnvironment'
 import { getComponent, hasComponent, UUIDComponent } from '@ir-engine/ecs'
 import { EditorComponentType } from '@ir-engine/editor/src/components/properties/Util'
 import { EditorControlFunctions } from '@ir-engine/editor/src/functions/EditorControlFunctions'
 import NodeEditor from '@ir-engine/editor/src/panels/properties/common/NodeEditor'
-import { GrabbableComponent } from '@ir-engine/engine/src/interaction/components/GrabbableComponent'
+import { GrabbableComponent } from '@ir-engine/engine/src/grabbable/GrabbableComponent'
 import { InteractableComponent } from '@ir-engine/engine/src/interaction/components/InteractableComponent'
-import { grabbableInteractMessage } from '@ir-engine/engine/src/interaction/functions/grabbableFunctions'
 import { GiGrab } from 'react-icons/gi'
+
+export const grabbableInteractMessage = 'Grab'
 
 export const GrabbableComponentNodeEditor: EditorComponentType = (props) => {
   const { t } = useTranslation()
 
   useEffect(() => {
-    if (isClient) {
-      if (!hasComponent(props.entity, InteractableComponent)) {
-        EditorControlFunctions.addOrRemoveComponent([props.entity], InteractableComponent, true, {
-          label: grabbableInteractMessage,
-          callbacks: [
-            {
-              callbackID: GrabbableComponent.grabbableCallbackName,
-              target: getComponent(props.entity, UUIDComponent)
-            }
-          ]
-        })
-      }
+    if (!hasComponent(props.entity, InteractableComponent)) {
+      EditorControlFunctions.addOrRemoveComponent([props.entity], InteractableComponent, true, {
+        label: grabbableInteractMessage,
+        callbacks: [
+          {
+            callbackID: GrabbableComponent.grabbableCallbackName,
+            target: getComponent(props.entity, UUIDComponent)
+          }
+        ]
+      })
     }
   }, [])
